First partially working firmware
I was able to make a custom docker image for esp-idf 4.0.1 and then use that to compile the tensorflow microlite library using the esp32 extensa cross compiler.

Then I use the update-microlite.sh script to move the library over into the lib directory.

Next I was able to make a C and C++ micropython module called microlite that builds to a flashable firmware.

I haven't completed the connection but expect to be able to make a custom class to hold the tensorflow interpreter and model data.