Add micropython lib/berkeley-db-1.xx patch
For some reason with the esp-idf compiler in the esp-idf 4.0.1 Docker
image it didn't like the include file format.

So I needed to wrap the inclue files with an actual #include "" on the
file names.

This patch needs to be applied before micropython will build to
micropython/lib/berkeley-db-1.xx