Callback functionality works
I am continuing to work on externalizing functionality to the python side so that the model and logic for setting up
the input and output tensors can be placed there.

I have the callbacks being passed through on the constructor of the interpreter object.  But because the TfLiteTensor is not a
micropython type then trying to set data from the python side is not working.

I think I will be able to either hide the tensor data in the interpreter class instance or pass it through to new helper methods
added to the interpreter.

These can then be called in the callbacks to setup the tensor appropriately for the model being run.