Inference is working
I needed to add the a tensorflow build flag to the options so that visual studio would use the
correct values for debugging.

I also had forgotten to add the type object for the tensor object.

Once visual studio was using the correct header files I was able to debug through and then the
hello world model showed up properly as an int8 model.

I just added the quantization methods and it works.