Changed micropython upstream version
We are based on the miketeachman i2s branch and he updated his code so I
needed to rebase my branch onto his latest work to get the i2s examples
to work which is needed for the microspeech example.

Minor changes to the hello-world.py file to facilitate running on the
esp32 board.

By putting the inference logic into a run method it allowed the it to be
called from a main.py on the esp32.