Add new audio_frontend module
For the micro_speech example we need to bundle the audio frontend code from tensorflow's experimental microfrontend which is responsible for converting the wav data into spectrogram that can be processed by the input tensor.

Move the makefiles up a level since they are not microlite specific.

Add docker file that adds xxd to the base 4.0.1 esp-idf image.

Add start of the example micropython script files for the micro_speech example.