Microspeech working in unix for 1 second data arrays
I took the example 1 second yes and no examples and setup the main-unix.py script so that
it will split up each 16000 pcm source array into 49 480 byte arrays.

The last one is 320 bytes and not working right yet.

My model was out of date so switching to the latest changed the type from uint8 to int8 and that seems to
have fixed inteference.

I may need to scale the output but it does have a positive value in the category.