Extract micro_speech module from esp32 main.
We seem to be able to sample audio when running on esp32 but the
inference isn't working right.

I changed the i2s_debug.py program to write two files to flash:
1. a wav playable file
2. a pcm (sampled i2s bytes without wav header)

I also wonder if the slicing mechanism is buggy.  To find out I have
made a new lib/micro_speech.py module to handle the slicing in common
between the unix and esp32 version.

I should be able to slice the original yes and no example data aswell as
slice the sampled audio in the debugger to make sure it is accurate.

Another possibility is that the dma to micropython audio buffers are
sized wrong an we are dropping audio packets.