Add 16khz and 8 khz example data
I used the esp32/i2s_dump.py file to sample gapless 2 second audio
samples for yes and no and 'one two three'.

Now to see if these inference properly with the unix example. and also
the training script to see if the spectrograms look correct and work to
classify these audio samples.