Create file in lib directory to force it to exist for the ci
It looks like the build of the tensorflow lite static library is working except the lib directory doesn't
exist because there are no files present within it.

A quick fix is to make a file within the lib directory which will cause it to exist and cause the
libtensorflow-microlite.a to be created with the correct name within the ci build.

Add more debugging to see where the firmware is generated so we can identify the path to archive.