Add person_detection example file
This bootstraps the model and then loads the no person image and then the person image and classifies both:

~/git/tensorflow-micropython-examples/examples/person_detection$ ./micropython ./person_detection.py
interpreter_make_new: model size = 300568, tensor area = 139264
Classify No Person Image
setup 9612 bytes on the inputTensor.
'not a person' = -25, 'person' = 25
Classify Person Image
setup 9612 bytes on the inputTensor.
'not a person' = -113, 'person' = 113

I need to verify the output with the c++ version to see if the outputs are what we expect.