Setup configurable ops
To start with put the libtf-op-resolvers.o into the
libtensorflow-microlite.a but its possible it should just be part of the
firmware.

Through testing it doesn't appear that the dead-code removal takes place
within the building of the static library.

Instead I suspect it will be done as part of the firmware build.