Experimental work to get to a build that can be debugged
At this point I think I should be able to use the bare-metal-gc-heap.c
functions to allow supporting malloc and free using the gc heap and then
use this for the C and C++ parts of tensorflow.

I also added back the alignment logic but it didn't seem to make any
difference on the board.