fix audio_frontend loadprohibited errors
I was able to get setup in WSL to be able to debug using openocd from windows and find the cause of the LoadProhibited
error that was being encountered whenever the microlite.audior_frontend was printed.

There was some missing configuration when the type was created.  the linkage to the type configuration for the class which
includes the reference to the class that is going to be used for printing the object.