stm32: add debug launcher to openocd in windows
When you run visual studio code in windows subsystem for linux and run openocd on the windows side it can work.

The IP in the launcher comes from the IP you see in /etc/resolve.conf

The openocd in windows needs a -c "bindto 0.0.0.0"  option and the firewall needs to be disabled on the hyperv virtual
network card.

Can be disabled in powershell as:
 Set-NetFirewallProfile -DisabledInterfaceAliases "vEthernet (WSL)"