stm32: enable cmsis build and fix firmware
This is for #33.

The stm32lib/bsearch.c was picked up from the latest openmv/tensorflow-lib repo.

The other stm32lib files were taken from newlib.

The approach taken here is to link the full math library instead of porting each operation
needed by tensorflow but not included out of the box.

The CMSIS version in micropython/lib/cmsis/inc are not aligned with what tensorflow downloads.

The prepare-tflm-stm32.sh script has been modified so that it will update micropython to the latest
before we build our custom firmware.