unix: fix build (#39)
I was able to identify if its the unix build in the microlite GNU Makefile.  Using this I set the board
to "unix" and then made some changes so that for unix

it will use the non-cmsis sources very similiar to how it works for ESP32.

I removed the old scripts which are deprecated.

This built locally but I'm not 100% on the github actions build script.