tflm: make ops configurable
This updates the make ops configurable work and applies it to trying to
support building a firmware for stm32 boards with only 512 KB flash.

The default is for all ops to be on.

A board can be customized to reduce the ops enabled.

That being said it doesn't quite work right.

I was unable to get link time optimization to work right so needed to take
precise control over which files are compiled.

This breaks with how tflm expects to work so there are linking errors.

But I hope so long as the linking error is for methods we are disabling then
it should be fine.