esp32-cam: adjustments to integrate camera image acquistion work
Renamed the board config to MICROLITE_SPIRAM_CAM.  It also works with my m5 timer camera which has 4MB Flash, 8MB PSRAM
and OV3660 camera.

I reshaped how the board heap is configured in micropython and the patch to apply those changes.  The board config
and headers are setup to use them.