stm32: Change relative path for brining ulab into the build.
In the previous commit I moved the block that does the symbolic link to the ulab code into
microlite-modules into the mpy cross compiler step.

It looks like there is a relative path issue due to moving the command.  This commit tries to account
for running it in a different directory.