Remove dynamic memory allocation and switch back to using stack based approach from openmv
Instead of using the new keyword use the stack based approach used by openmv.

Instead of newing the tflite::MicroInterpreter store it in the stack and then within the invoke method reallocate the tensors
and invoke the callback functions using the input and output tensors which are live within that nested context.