Remove git submodules in favour of shallow checkout approach
Instead of using submodules create a script that will make a shallow git clone of the dependency repos:
1. ulab
2. micropython
3. tflite-micro-esp-examples
4. tflite-micro (for non esp32 boards)

Pin the versions to the latest releases of each dependency.

ulab: 6.0.12
micropython: master
esp32-tflite: v1.0.0
tflite-micro: <undefined>

tflite-micro doesn't tag releases so its undefined to be the latest commit on the tip of the main branch.

use latest micropython to get upgrade to idf 5.0.2